+++ scripts/compare-subql-data/.mocharc.js
@@ -1,0 +1,4 @@
+module.exports = {
+  timeout: 500000,
+  require: ['ts-node/register/transpile-only']
+};
