+++ scripts/compare-subql-data/package.json
@@ -5,19 +5,24 @@
   "license": "MIT",
   "scripts": {
     "start": "ts-node --transpile-only src/index.ts",
-    "dev": "ts-node-dev --respawn --transpile-only src/index.ts"
+    "dev": "ts-node-dev --respawn --transpile-only src/index.ts",
+    "test": "mocha src/__tests__/*",
+    "lint": "eslint src"
   },
   "devDependencies": {
+    "@types/chai": "^4.3.4",
+    "@types/mocha": "^10.0.1",
     "@types/node": "^18.11.18",
     "@types/yargs": "^17.0.19",
     "@typescript-eslint/eslint-plugin": "^5.48.0",
     "@typescript-eslint/parser": "^5.48.0",
+    "chai": "^4.3.7",
     "csv-parse": "^5.3.3",
     "eslint": "^8.31.0",
+    "mocha": "^10.2.0",
     "ts-node": "^10.9.1",
     "ts-node-dev": "^2.0.0",
     "typescript": "^4.9.4",
     "yargs": "^17.6.2"
-  },
-  "dependencies": {}
+  }
 }
